requirements/ci.txt
# Linters
pep8
pep257
pylint

# Testing
pytest
pytest-describe
pytest-expecter
pytest-cov
pytest-random
scripttest == 1.3

# Coverage
coverage
coverage.space >= 0.6.1
